make(adviceCard)
separate advice preview card at component
