+++ src/services/adviceApi.jsx
@@ -13,9 +13,9 @@ const BASE_ENDPOINT = "/advices";
  * Gets all advices
  * @returns {Promise<Array>} List of advices
  */
-export async function getAllAdvices({ page = 1 } = {}) {
+export async function getAllAdvices() {
   try {
-    const response = await api.get(BASE_ENDPOINT, { params: { page } });
+    const response = await api.get(BASE_ENDPOINT);
     return handleApiResponse(response, "No advices found");
   } catch (error) {
     return handleApiError(error, "fetching advices");
