+++ src/utils/formatDateUtils.js
@@ -1,0 +1,5 @@
+export function formatDate(dateString) {
+  if (!dateString) return "";
+  const options = { year: "numeric", month: "long", day: "numeric" };
+  return new Date(dateString).toLocaleDateString(undefined, options);
+}
